Python Print function, String and Variables.
The print() is used to print a desired value in the output screen. The parameter in the print() function is of the data type string.
If we have to print an evaluated integer value, TYPECASTING has to be performed inside print(). This applies for all data types other than string. 

Ways to solve complex equations are also explained using codes in the code cell below.

The print() can be used in different ways. Observe each of the code statement below to get the best understanding of it. 

In order to continue printing in the same line, use *print(" < statement> ", end=" ")* in the preceding print command.

The input() is used to TAKE USER INPUT.
By default, string is the data type when we use input().
In order to print values other than string, TYPECASTING is necessary inside the print() statement parameter.