Defining functions
A function is a block of code which only runs when it is called.
You can pass data, known as parameters, into a function.
A function can return data as a result.

Keyword used :- def

Syntax :- 

def < function_name>(parameter):

  {

  //task or commands


  }

**Arbitrary arguments**- (args)

If we do not know how many arguments that will be passed into our function, add a * before the parameter name in the function definition.

This way the function will receive a tuple of arguments, and can access the items accordingly.

**Keyword arguments -** (kwargs)

We can also send parameter as a key=value syntax similar to what we did in python dictionaries. The order of arguments will not matter.

**Arbitrary keyword arguments -**

If the number of keyword arguments is unknown, add a double ** before the parameter name.

return keyword- To let a function return a value, use the return statement.

pass keyword- function definitions cannot be empty, but if for some reason we have a function definition with no content, put in the pass statement to avoid getting an error.

**Recursion function :-**

A function which calls itself many times until a terminating condition is met, is called as a recursive function. Follow the example below to see the working of a recursive function.

**Lambda Function :-**

A lambda function is a small anonymous function which can take any number of arguments, but can only have one expression.

Syntax- lambda arguments : expression