Importing modules
Modules are Python code libraries which we can include in our project.A module is a file containing a set of functions that we can include in our code for performing a task in an application.

In Topic 2, we saw the use of math() which is an inbuilt function.

ASSUME math() was NOT an inbuilt function. In that case we would have to first import the math module and then work on it. Look at the code below to see the working.

Syntax- import < module_name>

**Creating modules-**

1. Write and define a function with proper function definition.
2. Save that file having the function as file_name.py
3. import the file_name in your code in order to use the function which is defined in our application.

**Re-Naming a module-**

A module can be imported and renamed by using the 'as' keyword.

Syntax- import < module_name> as < new_name>

**dir() function-**
In order to see all the available functions in the module, use dir().

**Import from module-**

This is used to use an object which has been already defined and used in a module. From an example at www.w3schools.com, we consider the following case. We can't run it here as we have not saved the mymodule.py file

def greeting(name): #here we are creating a module named mymodule.py
  
  print("Hello, " + name)

person1 = {"name": "John","age": 36,"country": "Norway"} //an object used in greeting() module.

from mymodule import person1 //Importing person1 from the module

print("By importing person1, we get the following result")

print (person1["age"])