Downloading any external library module using pip
pip is a package manager for python packages or modules.

New modules can be installed using pip. There are many modules which are not found inbuilt in python-3. pip in pre-installed in python 3.4 and above.

Move to your command prompt to your Python script directory in order to work with pip.

Download package using pip-

Syntax :- pip install < package_name>

Note: To work with pip, we have to go to command prompt. Here in the code cell we can one pip command at a time. So the output of each pip command is printed at the output window of the code cells below.

Uninstalling a package-

Syntax- pip uninstall < package_name>

Once the command is run, it will ask whether we are sure that we want to uninstall the package. Reply with y/n as required.